+++ src/eulertools/subcommands/test.py
@@ -37,7 +37,7 @@ class Test:
         for language, problem, summary in runner.get_summaries(
             self.languages, self.problems
         ):
-            if not summary.problem_successful(language, problem):
+            if not summary.success(language, problem):
                 self.success = False
             self.print_summary(language, problem, summary)
         if not self.success: